refact(custom): extract common helpers for transition
