+++ src/chart/custom/CustomSeries.ts
@@ -43,7 +43,7 @@ import {
     TextCommonOption,
     ZRStyleProps
 } from '../../util/types';
-import Element, { ElementProps } from 'zrender/src/Element';
+import Element from 'zrender/src/Element';
 import SeriesData, { DefaultDataVisual } from '../../data/SeriesData';
 import GlobalModel from '../../model/Global';
 import createSeriesData from '../helper/createSeriesData';
@@ -64,24 +64,15 @@ import {
     Sector
 } from '../../util/graphic';
 import { TextStyleProps } from 'zrender/src/graphic/Text';
-
-
-export interface LooseElementProps extends ElementProps {
-    style?: ZRStyleProps;
-    shape?: Dictionary<unknown>;
-}
+import {
+    ElementTransformTransitionOptionMixin,
+    ElementTransitionOptionMixin,
+    TransitionBaseDuringAPI,
+    TransitionDuringAPI
+} from '../../animation/customGraphicTransitionHelper';
+import { TransformProp } from 'zrender/lib/core/Transformable';
 
 export type CustomExtraElementInfo = Dictionary<unknown>;
-export const TRANSFORM_PROPS = {
-    x: 1,
-    y: 1,
-    scaleX: 1,
-    scaleY: 1,
-    originX: 1,
-    originY: 1,
-    rotation: 1
-} as const;
-export type TransformProp = keyof typeof TRANSFORM_PROPS;
 
 // Also compat with ec4, where
 // `visual('color') visual('borderColor')` is supported.
@@ -102,19 +93,7 @@ export const NON_STYLE_VISUAL_PROPS = {
 } as const;
 export type NonStyleVisualProps = keyof typeof NON_STYLE_VISUAL_PROPS;
 
-// Do not declare "Dictionary" in TransitionAnyOption to restrict the type check.
-export type TransitionAnyOption = {
-    transition?: TransitionAnyProps;
-    enterFrom?: Dictionary<unknown>;
-    leaveTo?: Dictionary<unknown>;
-};
-type TransitionAnyProps = string | string[];
-type TransitionTransformOption = {
-    transition?: ElementRootTransitionProp | ElementRootTransitionProp[];
-    enterFrom?: Dictionary<unknown>;
-    leaveTo?: Dictionary<unknown>;
-};
-type ElementRootTransitionProp = TransformProp | 'shape' | 'extra' | 'style';
+// Do not declare "Dictionary" in ElementTransitionOptions to restrict the type check.
 type ShapeMorphingOption = {
     /**
      * If do shape morphing animation when type is changed.
@@ -123,27 +102,9 @@ type ShapeMorphingOption = {
     morph?: boolean
 };
 
-export interface CustomBaseDuringAPI {
-    // Usually other props do not need to be changed in animation during.
-    setTransform(key: TransformProp, val: number): this
-    getTransform(key: TransformProp): number;
-    setExtra(key: string, val: unknown): this
-    getExtra(key: string): unknown
-}
-export interface CustomDuringAPI<
-    StyleOpt extends any = any,
-    ShapeOpt extends any = any
-> extends CustomBaseDuringAPI {
-    setShape<T extends keyof ShapeOpt>(key: T, val: ShapeOpt[T]): this;
-    getShape<T extends keyof ShapeOpt>(key: T): ShapeOpt[T];
-    setStyle<T extends keyof StyleOpt>(key: T, val: StyleOpt[T]): this
-    getStyle<T extends keyof StyleOpt>(key: T): StyleOpt[T];
-};
-
-
 export interface CustomBaseElementOption extends Partial<Pick<
     Element, TransformProp | 'silent' | 'ignore' | 'textConfig'
->>, TransitionTransformOption {
+>>, ElementTransformTransitionOptionMixin {
     // element type, required.
     type: string;
     id?: string;
@@ -155,16 +116,16 @@ export interface CustomBaseElementOption extends Partial<Pick<
     // `false` means remove the clipPath
     clipPath?: CustomBaseZRPathOption | false;
     // `extra` can be set in any el option for custom prop for annimation duration.
-    extra?: Dictionary<unknown> & TransitionAnyOption;
+    extra?: Dictionary<unknown> & ElementTransitionOptionMixin;
     // updateDuringAnimation
-    during?(params: CustomBaseDuringAPI): void;
+    during?(params: TransitionBaseDuringAPI): void;
 };
 
 export interface CustomDisplayableOption extends CustomBaseElementOption, Partial<Pick<
     Displayable, 'zlevel' | 'z' | 'z2' | 'invisible'
 >> {
-    style?: ZRStyleProps & TransitionAnyOption;
-    during?(params: CustomDuringAPI): void;
+    style?: ZRStyleProps & ElementTransitionOptionMixin;
+    during?(params: TransitionDuringAPI): void;
     /**
      * @deprecated
      */
@@ -178,10 +139,10 @@ export interface CustomDisplayableOptionOnState extends Partial<Pick<
     Displayable, TransformProp | 'textConfig' | 'z2'
 >> {
     // `false` means remove emphasis trigger.
-    style?: (ZRStyleProps & TransitionAnyOption) | false;
+    style?: (ZRStyleProps & ElementTransitionOptionMixin) | false;
 
 
-    during?(params: CustomDuringAPI): void;
+    during?(params: TransitionDuringAPI): void;
 }
 export interface CustomGroupOption extends CustomBaseElementOption {
     type: 'group';
@@ -195,9 +156,9 @@ export interface CustomGroupOption extends CustomBaseElementOption {
 export interface CustomBaseZRPathOption<T extends PathProps['shape'] = PathProps['shape']>
     extends CustomDisplayableOption, ShapeMorphingOption {
     autoBatch?: boolean;
-    shape?: T & TransitionAnyOption;
+    shape?: T & ElementTransitionOptionMixin;
     style?: PathProps['style']
-    during?(params: CustomDuringAPI<PathStyleProps, T>): void;
+    during?(params: TransitionDuringAPI<PathStyleProps, T>): void;
 }
 
 interface BuiltinShapes {
@@ -240,22 +201,22 @@ export type CustomPathOption = CreateCustomBuitinPathOption<keyof BuiltinShapes>
     | CustomSVGPathOption;
 
 export interface CustomImageOptionOnState extends CustomDisplayableOptionOnState {
-    style?: ImageStyleProps & TransitionAnyOption;
+    style?: ImageStyleProps & ElementTransitionOptionMixin;
 }
 export interface CustomImageOption extends CustomDisplayableOption {
     type: 'image';
-    style?: ImageStyleProps & TransitionAnyOption;
+    style?: ImageStyleProps & ElementTransitionOptionMixin;
     emphasis?: CustomImageOptionOnState;
     blur?: CustomImageOptionOnState;
     select?: CustomImageOptionOnState;
 }
 
 export interface CustomTextOptionOnState extends CustomDisplayableOptionOnState {
-    style?: TextStyleProps & TransitionAnyOption;
+    style?: TextStyleProps & ElementTransitionOptionMixin;
 }
 export interface CustomTextOption extends CustomDisplayableOption {
     type: 'text';
-    style?: TextStyleProps & TransitionAnyOption;
+    style?: TextStyleProps & ElementTransitionOptionMixin;
     emphasis?: CustomTextOptionOnState;
     blur?: CustomTextOptionOnState;
     select?: CustomTextOptionOnState;
@@ -394,9 +355,7 @@ export const customInnerStore = makeInner<{
     customImagePath: CustomImageOption['style']['image'];
     // customText: string;
     txConZ2Set: number;
-    leaveToProps: ElementProps;
     option: CustomElementOption;
-    userDuring: CustomBaseElementOption['during'];
 }, Element>();
 
 export default class CustomSeriesModel extends SeriesModel<CustomSeriesOption> {
